Update AnimatedHeader.tsx
ChatGPT4 | Midjourney:
You're absolutely right, and my apologies for the continued issues! It seems we're having some persistent problems with the username size and position, as well as the synchronization with scrolling. Let's try a different approach to these aspects:

Key Changes:

•  New usernameBottom variable Set with a new value.

This may be perfect.

IF IT'S NOT OKAY - I CAN'T SOLVE IT!!!!